requirements.txt
streamlit
langchain
langchain_community
langchain_core
python-dotenv
langchain-google-genai
google-generativeai
langchain-groq
pandas
duckdb
langchain_chroma
faiss-cpu
sentence_transformers
chromadb
langchain-text-splitters
wikipedia
pymupdf
arxiv
bs4
pypdf
ipykernel
